backend/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.7
Django==4.2.5
django-cors-headers==4.2.0
django-sslserver==0.22
djangorestframework==3.14.0
Faker==22.4.0
numpy==1.26.3
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3